python-zenoh-eyes/pyproject.toml
[project]
name = "zenoh-eyes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "eclipse-zenoh>=1.5.0",
    "flatbuffers>=25.2.10",
]
